Add description to Directory.Build.props for RequestBuilder
A new <Description> element has been added to the Directory.Build.props file. This element provides a description for the RestSharp.RequestBuilder project, stating that it is a wrapper class that uses Fluent Syntax to create RestRequest objects. This addition enhances the project's metadata, aiding in documentation and package management.