Feature/extension merthods (#3)
* Add WithBuilder method and update README

Introduce WithBuilder extension method for RestRequest in RestSharp.RequestBuilder.Extensions namespace. Update README with new usage example. Minor formatting change in RestSharp.RequestBuilder.csproj.

* Add overloads to WithBuilder and new RequestBuilder ctor

Introduce new overloads for the WithBuilder extension method in
RestRequestExtensions to support resource as string or Uri with
an HTTP method. Add a new constructor to RequestBuilder to
initialize with a Uri and HTTP method, ensuring Uri is not null
and converting it to a string for internal use.